Move tracker to a standalone file
